course jpa 수정

+++ coursemos_spring_boot/src/main/java/com/example/demo/service/CourseServiceImpl.java
@@ -39,7 +39,7 @@ public class CourseServiceImpl implements CourseService{
 	}
 
 	@Override
-	public Course getCourseByCourseId(Long courseId) {
+	public Course getCourseByCourseId(int courseId) {
 		return courseDao.getCourseByCourseId(courseId);
 	}
 	@Override
